Moved classification logic in separate file
